+++ lib/core/index.ts
@@ -8,4 +8,5 @@ export * from "./layout";
 export * from "./mini";
 export * from "./page";
 export * from "./partial";
+export * from "./redirect";
 export * from "./shared";