+++ lib/helpers/index.ts
@@ -8,4 +8,5 @@ export * from "./each";
 export * from "./error";
 export * from "./html";
 export * from "./is-htmx";
+export * from "./redirect-to";
 export * from "./repeat";